Refactor metrics collection into separate package
Refactor the code to collect additional PostgreSQL metrics and organize the codebase into packages.

* **Metrics Collection**:
  - Move all Prometheus gauge definitions from `main.go` to a new package `metrics`.
  - Add new metrics for user connections, user queries, and database transaction rates.
  - Create a function `RegisterMetrics` in `pkg/metrics/metrics.go` to register all metrics.
  - Update the `collectMetrics` function to use the metrics from the `metrics` package.

* **Code Organization**:
  - Create a new package `metrics` to handle all metric definitions and registrations.
  - Create a new file `pkg/metrics/collector.go` to handle metric collection logic.
  - Remove all Prometheus gauge definitions and the `collectMetrics` function from `main.go`.
  - Import the `metrics` package and call `metrics.RegisterMetrics` in the `init` function in `main.go`.
  - Update the main function in `main.go` to use the `collectMetrics` function from the `metrics` package.

* **Documentation**:
  - Update `README.md` to document the new metrics collected by the application.
  - Add a section in `README.md` to reflect the new package structure and provide instructions for adding new metrics.
