+++ pull_request_template.md
@@ -1,0 +1,24 @@
+# Pull Request Template
+
+## Description
+
+Please include a summary of the changes and the related issue. Please also include relevant motivation and context. List any dependencies that are required for this change.
+
+Fixes # (issue)
+
+## Changes Made
+
+- [ ] Added new metrics for user connections, user queries, and database transaction rates.
+- [ ] Updated documentation to reflect the new metrics.
+- [ ] Refactored code to improve performance and maintainability.
+
+## Checklist
+
+- [ ] My code follows the style guidelines of this project.
+- [ ] I have performed a self-review of my own code.
+- [ ] I have commented my code, particularly in hard-to-understand areas.
+- [ ] I have made corresponding changes to the documentation.
+- [ ] My changes generate no new warnings.
+- [ ] I have added tests that prove my fix is effective or that my feature works.
+- [ ] New and existing unit tests pass locally with my changes.
+- [ ] Any dependent changes have been merged and published in downstream modules.
